Add Login Functionality

+++ app.py
@@ -1,6 +1,7 @@
 #!/usr/bin/env python3
 
-from flask import Flask, render_template, request, redirect, url_for, make_response
+from re import L
+from flask import Flask, render_template, request, redirect, url_for, make_response, flash
 from dotenv import dotenv_values
 
 import pymongo
@@ -8,8 +9,19 @@ import datetime
 from bson.objectid import ObjectId
 import sys
 
+# modules for user authentication
+import flask
+import flask_login
+from werkzeug.security import generate_password_hash
+from werkzeug.security import check_password_hash
+
 # instantiate the app
 app = Flask(__name__)
+app.secret_key = 'secret'  # Change this!
+
+#set up flask-login for user authentication
+login_manager = flask_login.LoginManager()
+login_manager.init_app(app)
 
 # load credentials and configuration options from .env file
 # if you do not yet have a file named .env, make one based on the template in env.example
@@ -34,6 +46,41 @@ except Exception as e:
     print(' *', "Failed to connect to MongoDB at", config['MONGO_URI'])
     print('Database connection error:', e) # debug
 
+# class to represent user
+class User(flask_login.UserMixin):
+    def __init__(self,data):
+        self.id = data['_id'] # shortcut to _id field
+        self.data = data # all user data
+
+def locate_user(user_id=None, email=None):
+    '''
+    Return a User object for the user with the given id or email address, or None if no such user exists.
+    @param user_id: the user_id of the user to locate
+    @param email: the email address of the user to locate
+    '''
+    if user_id:
+        criteria = {"_id":ObjectId(user_id)}
+    else:
+        criteria = {"email":email}
+    
+    doc = db.user.find_one(criteria) # find user w/ this email
+    # if user exists, create user object and return
+    if doc:
+        user = User(doc)
+        return user
+    #else
+    return None
+
+@login_manager.user_loader
+def user_loader(user_id):
+    #   This function is called automatically by flask-login with every request the browser makes to the server.
+    return locate_user(user_id=user_id)
+
+# make the currently-logged-in user, if any, available to all templates as 'user'
+@app.context_processor
+def inject_user():
+    # print(flask_login.current_user)
+    return dict(user=flask_login.current_user)
 
 userId = "6351acad640dc9083d534403"
 
@@ -41,8 +88,10 @@ userId = "6351acad640dc9083d534403"
 
 @app.route('/deadline') # show deadlines
 def show_deadline():
+    if not flask_login.current_user.is_authenticated:
+        return render_template('login.html')
     today = datetime.datetime.today() - datetime.timedelta(days = 1)
-    docs = db.deadline.find({"user": ObjectId(userId), "due":{"$gte":today}}).sort("due", 1) # sort in descending order of created_at timestamp
+    docs = db.deadline.find({"user": ObjectId(flask_login.current_user.data['_id']), "due":{"$gte":today}}).sort("due", 1) # sort in descending order of created_at timestamp
     deadline = list(docs)
     for i in deadline:
         i["countdown"] = i["due"] - today
@@ -51,12 +100,16 @@ def show_deadline():
 
 @app.route('/deadline/add') # add deadlines
 def add_deadline():
+    if not flask_login.current_user.is_authenticated:
+        return render_template('login.html')
     today = str(datetime.datetime.today())
     today= today[:10]+"T"+today[11:16]
     return render_template('add_deadline.html', today=today) # render the hone template
 
 @app.route('/deadline/add', methods=['POST'])# add deadlines confirm with post request
 def submit_deadline():
+    if not flask_login.current_user.is_authenticated:
+        return render_template('login.html')
     title = request.form['dtitle']
     due = request.form['dtime'].replace('T', '-').replace(':', '-').split('-')
     due = [int(v) for v in due]
@@ -65,15 +118,17 @@ def submit_deadline():
         "title": title, 
         "due": datetime.datetime(*due),
         "priority": priority,
-        "user": ObjectId(userId)
+        "user": ObjectId(flask_login.current_user.data['_id'])
     }
     db.deadline.insert_one(doc)
     return redirect(url_for('show_deadline'))
 
 @app.route('/deadline/edit') # edit deadlines page
 def edit_deadline():
+    if not flask_login.current_user.is_authenticated:
+        return render_template('login.html')
     today = datetime.datetime.today() - datetime.timedelta(days = 1)
-    docs = db.deadline.find({"user": ObjectId(userId), "due":{"$gte":today}}).sort("due", 1) # sort in descending order of created_at timestamp
+    docs = db.deadline.find({"user": ObjectId(flask_login.current_user.data['_id']), "due":{"$gte":today}}).sort("due", 1) # sort in descending order of created_at timestamp
     deadline = list(docs)
     for i in deadline:
         i["countdown"] = i["due"] - today
@@ -82,11 +137,15 @@ def edit_deadline():
 
 @app.route('/deadline/delete/<mongoid>')
 def delete_deadline(mongoid):
+    if not flask_login.current_user.is_authenticated:
+        return render_template('login.html')
     db.deadline.delete_one({"_id": ObjectId(mongoid)})
     return redirect(url_for('edit_deadline')) # tell the web browser to make a request for the / route (the home function)
 
 @app.route('/deadline/edit/<mongoid>')
 def rewrite_deadline(mongoid):
+    if not flask_login.current_user.is_authenticated:
+        return render_template('login.html')
     doc = db.deadline.find_one({"_id": ObjectId(mongoid)})
     time = str(doc["due"])
     time= time[:10]+"T"+time[11:16]
@@ -96,6 +155,8 @@ def rewrite_deadline(mongoid):
 
 @app.route('/deadline/edit/<mongoid>', methods=['POST'])# add deadlines confirm with post request
 def submit_edit_deadline(mongoid):
+    if not flask_login.current_user.is_authenticated:
+        return render_template('login.html')
     title = request.form['dtitle']
     due = request.form['dtime'].replace('T', '-').replace(':', '-').split('-')
     due = [int(v) for v in due]
@@ -104,7 +165,7 @@ def submit_edit_deadline(mongoid):
         "title": title, 
         "due": datetime.datetime(*due),
         "priority": priority,
-        "user": ObjectId(userId)
+        "user": ObjectId(flask_login.current_user.data['_id'])
     }
     db.deadline.update_one(
         {"_id": ObjectId(mongoid)}, # match criteria
@@ -130,17 +191,82 @@ def searching_deadline():
 
 @app.route('/account')
 def show_account():
-    return render_template('base.html') 
+    if not flask_login.current_user.is_authenticated:
+        return render_template('login.html')
+    return render_template('account.html', doc=flask_login.current_user.data)
+
+    # return render_template('account.html',author="asdf", docs="asdf")
+
+# route to show signup
+@app.route('/signup',methods=['GET'])
+def signup():
+    if flask_login.current_user.is_authenticated:
+        flash('You are already logged in.')
+        return redirect(url_for('todo.html'))
+    #else
+    return render_template('signup.html')
+
+# route to  handle the submission of the login form
+@app.route('/signup',methods=['POST'])
+def signup_submit():
+    email = request.form['email']
+    password = request.form['password']
+    hashed_password = generate_password_hash(password)
+    print('here')
+    if locate_user(email=email):
+        flash('An account for {} already exists. Please log in.'.format(email))
+        return redirect(url_for('login'))
+    user_id = db.user.insert_one({"email":email, "password": hashed_password}).inserted_id
+    if user_id:
+        user = User({
+            "_id":user_id,
+            "email":email,
+            "password":hashed_password
+        })
+        flask_login.login_user(user)
+        flash('Thanks for joining, {}!'.format(user.data['email']))
+        return redirect(url_for('add_task_home'))
+
+    return 'Signup failed'
+
+@app.route('/login',methods=['GET'])
+def login():
+    if flask_login.current_user.is_authenticated:
+        flash('You are already logged in!')
+        return redirect(url_for('show_deadline'))
+    return render_template('login.html', message='')
+
+@app.route('/logout')
+def logout():
+    flask_login.logout_user()
+    flash('You have been logged out.  Bye bye!')
+    return redirect(url_for('login'))
+
+@app.route('/login',methods=['POST'])
+def login_submit():
+    email = request.form['email']
+    password = request.form['password']
+    user = locate_user(email=email)
+    if user and check_password_hash(user.data['password'],password):
+        flask_login.login_user(user)
+        flash('Welcome back!')
+        return redirect(url_for('show_deadline'))
+        print(here)
+    return render_template('login.html', message = 'Wrong Password')
 
 #TO-DO Functions
 @app.route('/todo') #show to-do list
 def show_todo():
-    docs = db.todo.find({"user": ObjectId(userId)})
+    if not flask_login.current_user.is_authenticated:
+        return render_template('login.html')
+    docs = db.todo.find({"user": ObjectId(flask_login.current_user.data['_id'])})
     todo = list(docs)
     return render_template('todo.html', docs = todo)
     
 @app.route('/todo/complete/<mongoid>') #set checked tasks to completed in database
 def complete_task(mongoid):
+    if not flask_login.current_user.is_authenticated:
+        return render_template('login.html')
     doc = db.todo.find_one({"_id": ObjectId(mongoid)})
     status = not doc["completed"]
     db.todo.update({"_id": ObjectId(mongoid)},{"$set": {"completed": status}})
@@ -148,24 +274,31 @@ def complete_task(mongoid):
 
 @app.route('/todo/delete/<mongoid>')
 def delete_task(mongoid):
+    if not flask_login.current_user.is_authenticated:
+        return render_template('login.html')
     db.todo.delete_one({"_id": ObjectId(mongoid)})
     return redirect(url_for('edit_todo'))
 
 @app.route('/todo/edit')
 def edit_todo():
-    date = "today"
-    doc_list = db.todo.find({"user": ObjectId(userId)})
+    if not flask_login.current_user.is_authenticated:
+        return render_template('login.html')
+    doc_list = db.todo.find({"user": ObjectId(flask_login.current_user.data['_id'])})
     tasks = list(doc_list)
     return render_template('edit_task.html', docs=tasks)
 
 @app.route('/todo/edit/<mongoid>')    
 def rewrite_task(mongoid):
+    if not flask_login.current_user.is_authenticated:
+        return render_template('login.html')
     date = "today"
     doc = db.todo.find_one({"_id":ObjectId(mongoid)})
     return render_template('rewrite_todo.html',doc = doc, date = date)
 
 @app.route('/todo/edit/<mongoid>', methods=['POST']) #post request method
 def editing_existing_task(mongoid):
+    if not flask_login.current_user.is_authenticated:
+        return render_template('login.html')
     date = "today"
     title = request.form['ttitle']
     priority = request.form['tPriority']
@@ -177,7 +310,7 @@ def editing_existing_task(mongoid):
         "description": description,
         "label": label,
         "completed": False,
-        "user": ObjectId(userId)
+        "user": ObjectId(flask_login.current_user.data['_id'])
     }
     db.todo.update_one(
         {"_id": ObjectId(mongoid)},
@@ -187,12 +320,16 @@ def editing_existing_task(mongoid):
 
 @app.route('/todo/add') #add task screen w/o Post method
 def add_task_home():
+    if not flask_login.current_user.is_authenticated:
+        return render_template('login.html')
     today = str(datetime.datetime.today())
     today= today[:10]+"T"+today[11:16]
     return render_template('add_task.html', today=today) # render the hone template
 
 @app.route('/todo/add', methods=['POST']) # add task
 def add_task():
+    if not flask_login.current_user.is_authenticated:
+        return render_template('login.html')
     date = "today"
     title = request.form['ttitle']
     priority = request.form['tPriority']
@@ -204,7 +341,7 @@ def add_task():
         "description": description,
         "label": label,
         "completed": False,
-        "user": ObjectId(userId)
+        "user": ObjectId(flask_login.current_user.data['_id'])
     }
     db.todo.insert_one(newdoc)
     return redirect(url_for('show_todo'))
